+++ mingw-w64-libraries/winpthreads/src/rwlock.c
@@ -140,11 +140,11 @@ void rwl_print(volatile pthread_rwlock_t *rwl, char *txt)
     if (!print_state) return;
     rwlock_t *r = (rwlock_t *)*rwl;
     if (r == NULL) {
-        printf("RWL%p %d %s\n",(void *)*rwl,(int)GetCurrentThreadId(),txt);
+        printf("RWL%p %lu %s\n",(void *)*rwl,GetCurrentThreadId(),txt);
     } else {
-        printf("RWL%p %d V=%0X B=%d r=%ld w=%ld L=%p %s\n",
+        printf("RWL%p %lu V=%0X B=%d r=%ld w=%ld L=%p %s\n",
             (void *)*rwl,
-            (int)GetCurrentThreadId(), 
+            GetCurrentThreadId(),
             (int)r->valid, 
             (int)r->busy,
             0L,0L,NULL,txt);
